confine metabase-lib stuff to lib-adapter
