+++ frontend/src/metabase/embedding-sdk/lib/create-metabase-query/lib-adapter/clauses.ts
@@ -1,0 +1,3 @@
+export * from "./aggregations";
+export * from "./breakouts";
+export * from "./filters";
