--- frontend/src/metabase/embedding-sdk/lib/create-metabase-query/metabase-lib-query-clauses.ts
@@ -1,3 +1,0 @@
-export * from "./metabase-lib-query-aggregations";
-export * from "./metabase-lib-query-breakouts";
-export * from "./metabase-lib-query-filters";
